# docs/reference/openapi.kernel.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SoulOS Kernel",
    "version": "0.6.0"
  },
  "paths": {
...
    "/memory/ingest": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Ingest Memory",
        "operationId": "ingest_memory_memory_ingest_post",
        "parameters": [
          {
            "name": "X-SoulOS-Account-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Account-Id"
            }
          },
          {
            "name": "X-SoulOS-Gateway-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Gateway-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryIngest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/memory/retrieve": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Retrieve Memory",
        "operationId": "retrieve_memory_memory_retrieve_post",
        "parameters": [
          {
            "name": "X-SoulOS-Account-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Account-Id"
            }
          },
          {
            "name": "X-SoulOS-Gateway-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Gateway-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryRetrieve"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/memory/consolidate": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Consolidate Memory Route",
        "operationId": "consolidate_memory_route_memory_consolidate_post",
        "parameters": [
          {
            "name": "X-SoulOS-Account-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Account-Id"
            }
          },
          {
            "name": "X-SoulOS-Gateway-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Gateway-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryConsolidate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/memory/forget": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Forget Memory Route",
        "operationId": "forget_memory_route_memory_forget_post",
        "parameters": [
          {
            "name": "X-SoulOS-Account-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Account-Id"
            }
          },
          {
            "name": "X-SoulOS-Gateway-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Gateway-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryForget"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/memory/session/{bot_id}/{session_id}": {
      "delete": {
        "tags": [
          "memory"
        ],
        "summary": "Delete Session Memories Route",
        "operationId": "delete_session_memories_route_memory_session__bot_id___session_id__delete",
        "parameters": [
          {
            "name": "bot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bot Id"
            }
          },
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "X-SoulOS-Account-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Account-Id"
            }
          },
          {
            "name": "X-SoulOS-Gateway-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Gateway-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/memory/purge-expired": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Purge Expired Memories Route",
        "description": "Delete session-scoped memories and turn_sessions past MEMORY_SESSION_TTL_SECONDS.",
        "operationId": "purge_expired_memories_route_memory_purge_expired_post",
        "parameters": [
          {
            "name": "X-SoulOS-Account-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Account-Id"
            }
          },
          {
            "name": "X-SoulOS-Gateway-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Gateway-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryPurgeExpired"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/memory/sync": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Sync Memory",
        "operationId": "sync_memory_memory_sync_post",
        "parameters": [
          {
            "name": "X-SoulOS-Account-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Account-Id"
            }
          },
          {
            "name": "X-SoulOS-Gateway-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Gateway-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemorySync"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemoryConsolidate": {
        "properties": {
          "bot_id": {
            "type": "string",
            "title": "Bot Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 8
          }
        },
        "type": "object",
        "required": [
          "bot_id"
        ],
        "title": "MemoryConsolidate"
      },
      "MemoryForget": {
        "properties": {
          "bot_id": {
            "type": "string",
            "title": "Bot Id"
          },
          "content_match": {
            "type": "string",
            "maxLength": 32768,
            "title": "Content Match"
          }
        },
        "type": "object",
        "required": [
          "bot_id",
          "content_match"
        ],
        "title": "MemoryForget"
      },
      "MemoryIngest": {
        "properties": {
          "bot_id": {
            "type": "string",
            "title": "Bot Id"
          },
          "content": {
            "type": "string",
            "maxLength": 32768,
            "title": "Content"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "bot_id",
          "content"
        ],
        "title": "MemoryIngest"
      },
      "MemoryPurgeExpired": {
        "properties": {
          "bot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bot Id"
          }
        },
        "type": "object",
        "title": "MemoryPurgeExpired",
        "description": "Purge session-scoped memories past MEMORY_SESSION_TTL_SECONDS.\n\n``bot_id`` optional: omit to purge all bots (operator / CronJob; requires\nauth off or gateway without tenant account). When auth is enabled with an\naccount, ``bot_id`` is required."
      },
      "MemoryRetrieve": {
        "properties": {
          "bot_id": {
            "type": "string",
            "title": "Bot Id"
          },
          "query": {
            "type": "string",
            "maxLength": 32768,
            "title": "Query"
          },
          "top_k": {
            "type": "integer",
            "title": "Top K",
            "default": 5
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "include_scores": {
            "type": "boolean",
            "title": "Include Scores",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "bot_id",
          "query"
        ],
        "title": "MemoryRetrieve"
      },
      "MemorySync": {
        "properties": {
          "bot_id": {
            "type": "string",
            "title": "Bot Id"
          },
          "workspace_path": {
            "type": "string",
            "title": "Workspace Path"
          }
        },
        "type": "object",
        "required": [
          "bot_id",
          "workspace_path"
        ],
        "title": "MemorySync"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}